+++ plugins/blog/app/blog-view.ts
@@ -64,12 +64,14 @@ export class BlogView {
 
   delete(){
     var self = this;
-    this.client.delete('api/v1/blog/' + this.guid)
-      .then((response : any) => {
-        self.router.navigate(['/Blog', {filter: 'owner'}]);
-      })
-      .catch((e) => {
-      });
+    if(confirm("Are you sure?")){
+      this.client.delete('api/v1/blog/' + this.guid)
+        .then((response : any) => {
+          self.router.navigate(['/Blog', {filter: 'owner'}]);
+        })
+        .catch((e) => {
+        });
+    }
   }
 
 }
